+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Head
 
-- Fixed: `selector-type-no-unknown` now ignores nested selectors.
 - Fixed: `selector-pseudo-class-case`, `selector-pseudo-class-no-unknown`, `selector-pseudo-element-case`, `selector-pseudo-element-no-unknown` rules now ignore SCSS variable interpolation.
+- Fixed: `selector-type-no-unknown` now ignores nested selectors and keyframe selectors.
 
 # 6.4.1